requirements.txt
PyQt5~=5.15.11
twilio~=9.4.2
bcrypt~=4.2.1
PyOpenGL~=3.1.7
sounddevice~=0.5.1
soundfile~=0.13.0
numpy~=1.26.4
PyAudio~=0.2.14
pvporcupine~=3.0.3
SpeechRecognition~=3.13.0
PyYAML~=6.0.1
langchain~=0.3.14
pydantic~=2.11.5
langgraph~=0.2.72
langchain-community~=0.3.13
langchain-core~=0.3.35
langchain-openai~=0.3.1
tenacity~=8.5.0
yara-python~=4.5.1
psutil~=7.0.0
dnspython~=2.7.0
clamdpy~=0.1.0.post1
cryptography~=44.0.1
watchdog~=6.0.0
crewai~=0.102.0
requests~=2.32.3
openai~=1.73.0
colorama~=0.4.6
torch~=2.6.0
transformers~=4.46.1
inflect~=7.4.0
langdetect~=1.0.9
nltk~=3.9.1
translate~=3.6.1
SQLAlchemy~=2.0.36
peewee~=3.17.8
playhouse~=0.1.0
argparse~=1.4.0
pillow~=10.4.0
python-dotenv~=1.1.0
alembic~=1.14.0
litellm~=1.66.0
tokentrim~=0.1.13
yaspin~=3.0.2
toml~=0.10.2
tiktoken~=0.7.0
ipython~=8.31.0
html2image~=2.0.4.3
html2text~=2024.2.26
selenium~=4.24.0
webdriver-manager~=4.0.2
timm~=0.9.16
sentence-transformers~=3.4.1
ipykernel~=6.29.5
shortuuid~=1.0.13
janus~=1.0.0
uvicorn~=0.30.1
starlette~=0.46.2
fastapi~=0.111.0
rich~=13.9.4
packaging~=24.2
anthropic~=0.49.0
PyAutoGUI~=0.9.54
platformdirs~=4.2.0
e2b~=0.17.2a62
Send2Trash~=1.8.2
inquirer~=3.1.3
wget~=3.2
pyreadline3~=3.4.1
python-nmap~=0.7.1
opencv-python~=4.11.0.86
scipy~=1.15.2
schedule~=1.2.2
future~=1.0.0
librosa~=0.10.2.post1
netifaces~=0.11.0
easyocr~=1.7.1
yfinance~=0.2.52
setuptools~=77.0.3
langchain-text-splitters~=0.3.6
langchain-anthropic~=0.3.3
GitPython~=3.1.44
typer~=0.15.2
PyGithub~=2.6.1
pyfiglet~=0.8.post1
langchain-cohere~=0.3.4
inquirerpy~=0.3.4
aiohttp~=3.11.11
deep-translator~=1.11.4
matplotlib~=3.10.1
pandas~=2.2.3
pyperclip~=1.9.0
PubChemPy~=1.0.4
reportlab~=4.4.0
mendeleev~=1.0.0
chempy~=0.9.0
Wikipedia-API~=0.8.1
keyring~=25.6.0
webrtcvad~=2.0.10
plotly~=6.0.1
newspaper3k~=0.2.8
pandasai~=2.3.0
hassapi~=0.2.1
typing_extensions~=4.12.2
duckduckgo_search~=7.2.1
autogen~=0.8.6
pyautogen~=0.8.6
prompt_toolkit~=3.0.48
jupyter_client~=8.6.0
parler_tts~=0.2.3